Add and consolidate operations for resizing and rescaling rasters.
This checkin adds a Resize() operation for generating new rasters
with new extents and resolutions.  This is especially useful
when creating an image from an output raster (generated at the
resolution of your data) to a new extent and resolution specification.

It also adds a Rescale which will scale up or down the resolution of a
raster by a fixed percentage.

Resample() is now folded into Resize().
